Upload Experimental HLG LUT generator
I think this is all I can currently make by guessing without actual HDR device. I am asking others with HDR display to test this in proper HDR supported software (Resolve etc.), and modify this script to finish it. I am now writing down some of the things about HDR based on my understanding (note this is just my understanding, might not be super accurate):

1. One of my friends with HDR capable MacBook tested Blender 4.0's EDR/HDR support, it seems to be completely broken, both SDR and HDR are terribly broken in Blender 4.0 Mac version.  I guess this why Troy was freaking out about the EDR patch, but the official Blender devs don't seem to have an access to MacBook with a Built-in P3 EDR screen, therefore it might be hard for them to realize what is wrong. So maybe don't use Blender for testing this at the moment.

2. Don't trust those video players out there. I asked the friend I mentioned to play the same Rec.2100 PQ encoded video with HDR10 metadata embedded, he comfirmed that different video players each seems to include a unique view transform for HDR content. Including QuickTime, and iPhone 15 Pro's built-in video player. 
I don't yet know which APP to trust, maybe just use DaVinci Resolve for now.

3. HDR in terms of image formation, is weird.
If we think of our AgX Base Rec.2020 image as a percentage for a display (how many percent of emission strength this pixel should be, relative to that pixel's maximum emission strength), the max emission nits value specifically, doesn't seem to really matter, you can have a 1000 nits display, and we would multiply our percentage (0.0 to 1.0, or 0% to 100%) with that 1000 nits, and we get a nits based value for our image.

Therefore, if we follow a percentage based mentality for HDR, we can just do the exact same thing we have been doing all alone, that is multiplay the percentage by the max emission (1000 nits for HLG), and apply an HLG "[0 , 1000] nits to [0, 1] encoding" curve. Everything stays the same as our SDR image. 

However, this is not how most protocals are implemented. Most protocals like Resolve, TCAM, etc., seem to implement HDR encoding with "SDR = 100 nits" assumption. Basically, what we did above is kind of "SDR = 1000 nits", and with 100 nits assuption, we would calculate the ratio between SDR max (100) and HDR max (1000), and muliply the percentage by that ratio, basically in our case, scale the percentage down by 10. 

This is super super weird IMHO, if we think of percentage here again. Our image's 18% middle grey is now 1.8%.  So ultimately, comparing with SDR image, we are actually implementing a "darkening" in our image.

Note there has been controversies on the whole SDR 100 nits thing, as it was derived from a weird notion of "diffuse white" which doesn't exist in a formed image. Resolve also has a SDR = 203 nits implementation, Troy mentioned once that most actual SDR displays are around 300 nits. 

I am going with SDR = 100 nits currently, but there is a parameter in the script that you can change.

4. HDR image needs to match SDR image's appearance.

We cannot do it 100% matching since we don't have a perceptual color model that actually works right now (none of the models can describe the classic red train in an image that only has cyan pixels). But we still need to try our best to do the matching. 

For example, given a movie shot, a bunch of characters are looking at something in shock, the camera cuts to a glowing chromatic object, will it make sense that the HDR distribution of the movie can clearly see the chroma-laden details on the glowing object, while the SDR distribution of the movie only shows a solid block of white lights?

It doesn't make sense, therefore, SDR and HDR image need to match, at least in terms of attenuation's rate of change (I.E. the speed things fade to white). 

There are parameters in the script for tweaking the match. I don't have an HDR device so I cannot fine tune the match. Please, for people that pick this up from here, test this in Resolve and find a setting that matches the AgX Base Rec.2020 image. Or if none of the possible setting of the current parameters work, re-work the script to make it match.

The config for using this LUT can be found here:
https://github.com/EaryChow/AgX/tree/HDR-Experimental

With this published, I will probably no longer do any AgX-related development from now. At least I hope so, I need to focus on my rea-life situation, it's unhealthy for me to continue, but I keep finding myself coming back. Maybe it's because of the Inertia left from my devotion into this project during the past about 2 years. Hopefully I can actually stop developing this and focus on my life issues.