Upload AgX node Geometry Nodes demo file
This is a vision in which AgX is a dynamically tunable node in Blender's compositor. Though since Blender compositor doesn't have Matrix nodes yet, this demo file is constructed in Geometry Nodes.
The file is done with a Blender 4.5 alpha daily build. 
The node group includes the entirety of the codes that generated the AgX LUT used in Blender, including the dynamically tunable sigmoid, as well as the dynamically tunable Inset/Outset generator code (from SB2383's working_space.py) that I left out of this repo. This node group with different settings can become different variants of AgX, "Blender AgX", "Original AgX", "AgX SB2383", or "AgX Resolve DCTL Default" etc.. 

The node's UI is bigger than I expected before building it. Some thoughts:
- The working primaries and whitepoint inputs can ideally be an enum dropdown, with common options listed such as Rec.709, Display P3, Rec.2020 etc., maybe also include ARRI Wide Gamut 3 as an option for matching AgX Resolve's DCTL default. Then there should be a `Custom` option, by selecting which the four xyY input fields would got expanded out.
- The four Inset and Outset parameters may be all collapsed/folded by default.
This should prevent the node from looking too large.
But this is just a vision, there are no promises.