Use Panels on AgX_Procedural demo file UI
Arrange the input sockets of the node group into collapsible panels. This shrinks the visual size of the node by achieving what I mentioned in the commit message of the previous commit:
- The four Inset and Outset parameters may be all collapsed/folded by default.
But the working space primaries UI is not yet how I would imagine it to be, due to the fact that the "Menu" feature of the node group input is too limited at the moment. What I imagined it should be, as mentioned in the last commit:
- The working primaries and whitepoint inputs can ideally be an enum dropdown, with common options listed such as Rec.709, Display P3, Rec.2020 etc., maybe also include ARRI Wide Gamut 3 as an option for matching AgX Resolve's DCTL default. Then there should be a `Custom` option, by selecting which the four xyY input fields would got expanded out.