Fix weird precision issue in primary scaling
There is a weird precision issue in the arbitrary scaling of primary step, the python script outputs the scaled red as (1.6218190000000001, 0.3329310000000001), while doing the same thing in Blender nodes would yield (1.6218189, 0.33293104). The difference between the two sets of numbers are very small, they are arguably the same thing. But the numbers produced by the Blender nodes causes a weird bug when using the original AgX settings. I tried to copy and paste the python script number to the nodes, the nodes converts the precision to (1.62182, 0.332931), which then works fine. I then decided to snap the precision of the scaled primary to the increment of 0.00001, and this resulted in the number of (1.62181, 0.33293), the issue is now gone.